Add files via upload
Searches OMDB (http://www.omdbapi.com) for Movie details, specifically the TT Name for IMDB, based on the folder name that is assumed to represent a movie name with perhaps a year value.  Renames the folder to the TT name to improve the reliability of Kodi Universal Movie Scrapper to find movie details for all movie titles.